pyproject.toml
[project]
name = "django_DDD"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "lucasantonete@gmail.com", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=5.1.3",
    "pytest>=8.3.3",
    "pylint>=3.3.1",
    "djangorestframework>=3.15.2",
    "pytest-django>=4.9.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
